linking finetuning to inference  (#10)
