+++ Dockerfile
@@ -4,5 +4,6 @@ WORKDIR /training
 
 COPY . .
 
+RUN chmod 755 src/inference/serve.sh
 RUN pip install -r requirements.txt
-ENTRYPOINT ["python3", "src/scripts/finetune.py"]
+ENTRYPOINT ["python3", "src/finetune_inference_flow.py"]
