+++ requirements.txt
@@ -11,3 +11,6 @@ trl==0.7.1
 torch==2.0.1
 transformers==4.32.1
 s3fs==2023.9.0
+gradio==3.44.3
+typing_extensions==4.7.1
+fschat==0.2.28
