+++ src/akash/deploy.yml
@@ -3,10 +3,20 @@ version: "2.0"
 
 services:
   finetune:
-    image: tmontfort/finetune:latest
+    image: tmontfort/finetune:inference-v1.1
+    env:
+        - HUGGING_FACE_TOKEN=
+        - ACCESS_KEY_ID=
+        - SECRET_ACCESS_KEY=
+        - ENDPOINT_URL=
+    command:
+        - "sh"
+        - "-c"
+    args:
+        - 'python3 src/finetune_inference_flow.py --bucket_name <storj-bucket-name> --job_id <random-id>'
     expose:
-      - port: 8000
-        as: 80
+      - port: 7860
+        as: 8080
         to:
           - global: true
 
@@ -17,18 +27,23 @@ profiles:
         cpu:
           units: 8
         memory:
-          size: 50Gi
+          size: 16Gi
         gpu:
           units: 1
           attributes:
             vendor:
               nvidia:
+              - model: 3090
+              - model: v100
+              - model: a100
+              - model: h100
+              - model: a40
         storage:
           - size: 40Gi
   placement:
     akash:
       pricing:
-        falcon7b: 
+        finetune: 
           denom: uakt
           amount: 100000
